Death Knight Phase 1 Fixes
Revision #1
-Attempt to prevent the player-controlled Eye of Archerus from becoming
stuck pathing to a target when you press '1' or auto-attack until you
deselect your target. Attempted fix involves overriding the
AttackStart() function to do nothing.

-Currently you can duel multiple Death Knight Initiates at the same time
because you can start a duel with another target before the first one's
script checks if you're in combat before continuing the script.
Attempted fix involves checking if the player/NPC is in combat, and if
false will initiate combat with the player before they can interact with
another Initiate and start a second duel.

-The Death Knight Initiates currently do not cast spells and will simply
try to melee you to death... Copied the spell structure from the
Unworthy Initiates to the dueling Initiates.